adding in fixes for cancel behavior when editing economic units and devices

+++ Assets/Scripts/EuSavable.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 [System.Serializable]
 public class EuSavable {
-	public int latency;
+	public float latency;
 	public List<string> assets;
 	public List<int> qty;
 	public string euname;
